pyproject.toml
[project]
name = "srt-dataloader"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "sabinem", email = "5292683+sabinem@users.noreply.github.com" }
]
dependencies = [
    "typer>=0.12.5",
    "srt>=3.5.3",
    "python-dotenv>=1.0.1",
    "pymongo>=4.10.1",
]
readme = "README.md"
requires-python = ">= 3.11"

[project.scripts]
"srt-dataloader" = "srt_dataloader:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/srt_dataloader"]
